Fixed some compilation bugs
AMRData's C parameter once again defaults to 1. This functionality was temporarily removed to assist with debugging.

Stencil's devCoefs and devOffset methods are suddenly creating a problem for reasons unknown. Since they do not appear to be used any more, I have commented them out for the sake of cleaning up the API.