DataIterator Overhaul
DataIterator now depends only on DBLInternals and not on DisjointBoxLayout. This has been done to resolve a circular dependency between DataIterator and DisjointBoxLayout that was preventing the latter from implementing an STL complient iterator syntax. Said syntax requires DataIterator (and not DataIndex) to be the return type for begin() and end(). Changing this broke a massive amount of code due to the plethora of iter != layout.end() style checks in Proto. These checks have been corrected inside of Proto, but for convenience there are now == and != operators that allow DataIterator and DataIndex to be compared with eachother. As a nice upshot, DataIterable objects (DisjointBoxLayout and LevelBoxData) now support the very slick range based for-loop syntax (e.g. for (auto index : layout) where layout is a DisjointBoxLayout and index resolves to each DataIndex in the layout).

One casualty of this change is that the DataIterator::box() and DataIterator::point() functions have been removed. No functionality is lost here since DisjointBoxLayout::box(DataIterator) and DisjointBoxLayout::point(DataIterator) still exist. To be mildly pedantic, the removed functions never really made a ton of semantic sense.