Added LevelBoxData::copyTo test to CI
the copyTo test appears to be working correctly. Copying into a LBD with a larger ghost cell allocation than the source may result in the destination's extra ghost cells not being filled correctly. At the moment, this is not being considered a bug. Documentation has been added to LevelBoxData::copyTo which instructs users to call exchange manually in this case.