Templated DataIterator Refactor
DataIterator, DataIndex, and Copier are now templated on an object which defines the underlying data organization and has an interface for converting procerssor ID information into a global index space that can be iterated over. For data holders based on DisjointBoxLayout, the template object is BoxPartition

This seemingly insane refactor has been executed in anticipation of the mapped-multiblock interface which implements data on an array of patches which cannot be reduced to a single DisjointBoxLayout. The refactor will - hopefully - improve extensibility by recycling the code in DataIterator and - most notably - Copier

This refactor should be almost completely invisible to the end user.