Updates to LevelBoxData CI tests for debugging CUDA+MPI functionality
LevelBoxDataTests for exchange and copyTo are now separated for all HOST / DEVICE cases for better testing granularity. Currently the ExchangeDevice, CopyToDeviceToDevice, and CopyToHostToDevice tests are not passing with MPI+CUDA. CopyToDeviceToHost DOES pass interestingly, but this may be a bug in that test specifically.

Bugfix: HDF5::writeLevel will now copy patches one by one to the host when given a piece of DEVICE data. This strategy preserves the content of ghost cells which is not the generally case when LevelData::copyTo is called.